Add support for comparing body of the request alongside default compare
In this commit we add support for comparing the request body alongside the default
compare which only checked the request method as well as the URL string. We also updated
the tests to make them determinitic so that if there is no change in the request response cycle
(except for the 'Date' header) the cassette.Save() option doesn't overwrite the existing cassette
(this was a design decision we went with but if there is any problems we can easily change this behaviour).